+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 3.6.5
 - Debug command "-sjasmplus.path" added for dynamic loading of sld files during runtime (#156).
 - launch.json: Added "customVisualMemBlocks" (#157).
+- Fixed help text formatting
 
 # 3.6.4
 - launch.json: Added "load" support for .z80 files.